Update
-fix bug with data downloader
-add 2nd, much smaller model
-add flag for profiling
-add seed for data split

+++ data_util/ensure_data_downloaded.sh
@@ -37,6 +37,9 @@ fi
 
 
 # decompress puzzles
-rm data/lichess_db_puzzle.csv
+if [ -f data/lichess_db_puzzle.csv ]; then
+    rm data/lichess_db_puzzle.csv
+fi 
+
 bzip2 -kd data/lichess_db_puzzle.csv.bz2
